requirements.txt
pygooglenews
fastapi
uvicorn
python-dotenv
pytest
httpx 